move default mail client and startup settings to new general route
